Implementing Search Results - Part 2
In this commit I have implemented the load spinner in the .results (the place where should appear the searched recipes. Also the render for searched recipes is working after the implementation in the resultsView.js. And the creation of error msg propertie at resultsView.js similar with the one at recipeView.js. The error msg was set to render in the View.js. In general togheter with the last commit I implemented the event that the user searches, after the async loaded the search results and then render it to our user interface as the description of the flowchart-part-1