Implementing Pagination-Part 1
In this commit the pagination functionality was parcialy implemented. I have created a getSearchResultsPage function using the array state (line 6) data in the model.js. The parameters of the function stored in RES_PER_PAGE in the config.js was then exported to model.js. Finally, the function was stored and exported from the model.js to the controller.js. This function was called in the controlSearchResults in the line 53 of the controller.js.