Implementing Pagination-Part 2
In this commit work around to render buttons to change the searched recipes pages considering in the last commit I've set the search results to 10 results per page. A new file paginationView.js was created linking the prev and used buttons in the html. To compute which data regarding search, page and resultsPerPage we employ with the respective buttons we use the state array from model.js. In the controller we called the render function to pass the data that we need to be able to use it. The paginationView.js was exported and accessed inside the controller. The eventlistener were added to the buttons of the previous page and next page.