Updating Recipe Servings
In this commit a controlServings function was created in the controller.js. A new updateServings function was created in the model.js and it was called in the controller.js to update the servings. This function logic was to increase and decrease the amount of ingredients when the number of servings increase or decrease. An addEventListener was created in the recipeView.js with the name of addHandlerUpdateServings to listen to the buttons of decreasing and increasing servings. This was connected with the controller.js by recipeView.addHandlerUpdateServings() receiving the controlServings information: recipeView.addHandlerUpdateServings(controlServings). Finally, the handler() was called in the recipeView.js in the line 23 passing the updateTo.