Developing a DOM Updating Algorithm
In this commit I have developed an algorithm which will update the DOM only in places where it actually changed. This is necessary because as the servings are updated basically re-render the entire recipe view losing performance. Also, from the loaded recipes I implemented a mark on the selected pizza and when another pizza is selected it does not upload the images of all the pizzas.