Implementing Bookmarks - Part 1
In this commit, the first addBookmark function was created in the model.js and exported. Also in the model.js an empty bookmarks array was created inside the state (line 6). Inside the controller.js a new controller to add bookmark was created ControlAddBookmark calling for the addBookmark from model.js. Inside the recipeView.js addHandlerAddBookmark addeventlistener method was created and set to the btn--bookmark button and called in the controller.js. Now after a click in the btn--bookmark appears an object in the console with the bookmarks property set to true. Finally the element was adjusted to be updated in the controller.js. For deleting the bookmark a deleteBookmark function was created in the model.js and exported and called in the controller.js.