Implementing Bookmarks - Part 2
In this commit a bookmarksView.js as created implementing a BookmarksView function to show the bookmarks results. This file was exported and then imported by the controller and used in the controlAddBookmark function to render bookmarks. Then in the controller.js the bookmarsView was update in the recipeContainer function to mark only the current recipe in the bookmark.