Storing Bookmarks With localStorage
In this commit I've set the storage to happen in the model.js. Then, a persistBookmarks function was created in model.js and called in the addBookmark and deleteBookmark function. In this way the bookmarked recipes are properly stored in local storage. But when the website is uploaded the bookmarks in the website disappear. Nevertheless, the bookmarked recipes are stored in local storage even after the webpage reloads. To adjust this, I have set to get the bookmark out of local storage back in the code and render it back in the bookmark view by creating a init function in the model.js.